+++ Editor/ARCoreBuildProcessor.cs
@@ -8,6 +8,7 @@ using UnityEditor.Android;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEditor.Callbacks;
+using UnityEditor.XR.ARSubsystems;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.XR.ARCore;
@@ -16,7 +17,7 @@ using Diag = System.Diagnostics;
 
 namespace UnityEditor.XR.ARCore
 {
-    internal class ARCorePreprocessBuild : IPreprocessBuildWithReport
+    internal class ARCorePreprocessBuild : IPreprocessBuildWithReport, IPostprocessBuildWithReport
     {
         public int callbackOrder { get { return 0; } }
 
@@ -31,6 +32,17 @@ namespace UnityEditor.XR.ARCore
             EnsureOnlyOpenGLES3IsUsed();
             EnsureGradleIsUsed();
             BuildImageTrackingAssets();
+
+            BuildHelper.AddBackgroundShaderToProject(ARCoreCameraSubsystem.backgroundShaderName);
+        }
+
+        public void OnPostprocessBuild(BuildReport report)
+        {
+            if (report.summary.platform != BuildTarget.Android)
+                return;
+
+            BuildHelper.RemoveShaderFromProject(ARCoreCameraSubsystem.backgroundShaderName);
+            RemoveGeneratedStreamingAssets();
         }
 
         void EnsureGradleIsUsed()
@@ -210,16 +222,22 @@ namespace UnityEditor.XR.ARCore
                             overallProgress + progressPerLibrary * (numSteps - 1) / numSteps);
 
                         var packagePath = Path.GetFullPath("Packages/com.unity.xr.arcore");
-    #if UNITY_EDITOR_WIN
-                        string platformName = "Windows";
-                        string extension = ".exe";
-    #elif UNITY_EDITOR_OSX
-                        string platformName = "MacOS";
+
                         string extension = "";
+                        string platformName = "Undefined";
+    #if UNITY_EDITOR_WIN
+                        platformName = "Windows";
+                        extension = ".exe";
+    #elif UNITY_EDITOR_OSX
+                        platformName = "MacOS";
+                        extension = "";
+    #elif UNITY_EDITOR_LINUX
+                        platformName = "Linux";
+                        extension = "";
     #endif
                         var arcoreimgPath = Path.Combine(packagePath, "Tools~", platformName, "arcoreimg" + extension);
 
-    #if UNITY_EDITOR_OSX
+    #if UNITY_EDITOR_OSX || UNITY_EDITOR_LINUX
                         SetExecutablePermission(arcoreimgPath);
     #endif
 
@@ -296,13 +314,6 @@ namespace UnityEditor.XR.ARCore
             RemoveDirectoryWithMetafile(ARCoreImageTrackingProvider.k_StreamingAssetsPath);
             if (s_ShouldDeleteStreamingAssetsFolder)
                 RemoveDirectoryWithMetafile(Application.streamingAssetsPath);
-        }
-
-        [PostProcessBuild(1)]
-        public static void OnPostProcessBuild(BuildTarget target, string pathToBuiltProject)
-        {
-            if (target == BuildTarget.Android)
-                RemoveGeneratedStreamingAssets();
         }
 
         static bool s_ShouldDeleteStreamingAssetsFolder;