+++ Runtime/ARCoreLoaderConstants.cs
@@ -1,0 +1,17 @@
+#if UNITY_2019_2_OR_NEWER
+using System;
+
+namespace UnityEngine.XR.ARCore
+{
+    /// <summary>
+    /// Static constants
+    /// </summary>
+    public static class ARCoreLoaderConstants
+    {
+        /// <summary>
+        /// Key we use to store and retrieve custom configuration settings from EditorBuildSettings
+        /// </summary>
+        public const string k_SettingsKey = "com.unity.xr.management.arcore_loader";
+    }
+}
+#endif